orgabize fixtures and add option for default installer prefix
